+++ docs/docs/components/tabs.mdx
@@ -19,7 +19,7 @@ import { Tabs } from "@hover-design/react";
 import { useState } from "react";
 
 ```jsx
-const tabData = [
+const tabsData = [
   {
     label: "First",
     value: "first",
@@ -43,7 +43,7 @@ const tabData = [
 ];
 
 <div className="App">
-  <Tabs defaultValue={"first"} tabData={tabData}>
+  <Tabs defaultValue={"first"} tabsData={tabsData}>
     {(selectedTab) => <div>{selectedTab.label}</div>}
   </Tabs>
 </div>;
@@ -53,7 +53,7 @@ const tabData = [
 
 <Tabs
   defaultValue={"first"}
-  tabData={[
+  tabsData={[
     {
       label: "First",
       value: "first",
@@ -112,7 +112,7 @@ const tabData = [
 ```jsx
 const [activeTab, setActiveTab] = useState("first");
 
-const tabData = [
+const tabsData = [
   {
     label: "First",
     value: "first",
@@ -140,14 +140,14 @@ const tabData = [
     value={activeTab}
     onChange={(selectedTab) => setActiveTab(selectedTab.value)}
     value
-    tabData={tabData}
+    tabsData={tabsData}
   >
     {(selectedTab) => <div>{selectedTab.label}</div>}
   </Tabs>
 </div>;
 ```
 
-> **_(tabData)_** contains the whole tab object which is selected.
+> **_(tabsData)_** contains the whole tab object which is selected.
 
 export const App = () => {
   const [activeTab, setActiveTab] = useState("first");
@@ -156,7 +156,7 @@ export const App = () => {
       grow
       value={activeTab}
       onChange={(selectedTab) => setActiveTab(selectedTab.value)}
-      tabData={[
+      tabsData={[
         {
           label: "First",
           value: "first",
@@ -225,7 +225,7 @@ export const App = () => {
 | onChange     |                           `(selectedTabObject)=>void`                            |   `()=>{}`    |        Yes |
 | grow         |                      grows to fit the containter `boolean`                       |    `false`    |        Yes |
 | height       |                                     `string`                                     |    `40px`     |        Yes |
-| tabData      |                           `Array of` **_Tab Object_**                            |               |         No |
+| tabsData     |                           `Array of` **_Tab Object_**                            |               |         No |
 | children     |  _ not a prop but children to the `<Tabs/>`_ <br/> `(selectedTabObject)=> JSX `  |               |         No |
 
 ### Tabs Object
